requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
cffi==1.15.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
entrypoints==0.3
fonttools==4.28.3
h5py==3.6.0
imageio==2.13.3
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lumicks.pylake==0.11.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.21.4
opencv-python-headless==4.5.4.60
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
py==1.11.0
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
PyWavelets==1.2.0
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
scikit-image==0.19.0
scikit-learn==0.24.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
tabulate==0.8.6
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
